+++ tests/path/regex_filters.test.ts
@@ -4,7 +4,7 @@ import { Match, Search } from "../../src/path/functions";
 describe("match filter", () => {
   test("with caching", () => {
     const env = new JSONPathEnvironment();
-    env.filterRegister.set("match", new Match({ cacheSize: 10 }));
+    env.functionRegister.set("match", new Match({ cacheSize: 10 }));
     let rv = env.query("$[?match(@.a, 'a.*')]", [{ a: "ab" }]);
     expect(rv.values()).toStrictEqual([{ a: "ab" }]);
     rv = env.query("$[?match(@.a, 'a.*')]", [{ a: "ac" }]);
@@ -12,13 +12,13 @@ describe("match filter", () => {
   });
   test("without caching", () => {
     const env = new JSONPathEnvironment();
-    env.filterRegister.set("match", new Match({ cacheSize: 0 }));
+    env.functionRegister.set("match", new Match({ cacheSize: 0 }));
     const rv = env.query("$[?match(@.a, 'a.*')]", [{ a: "ab" }]);
     expect(rv.values()).toStrictEqual([{ a: "ab" }]);
   });
   test("throw error without caching", () => {
     const env = new JSONPathEnvironment();
-    env.filterRegister.set(
+    env.functionRegister.set(
       "match",
       new Match({ cacheSize: 0, throwErrors: true }),
     );
@@ -31,7 +31,7 @@ describe("match filter", () => {
 describe("search filter", () => {
   test("with caching", () => {
     const env = new JSONPathEnvironment();
-    env.filterRegister.set("search", new Search({ cacheSize: 10 }));
+    env.functionRegister.set("search", new Search({ cacheSize: 10 }));
     let rv = env.query("$[?search(@.a, 'a.*')]", [{ a: "the end is ab" }]);
     expect(rv.values()).toStrictEqual([{ a: "the end is ab" }]);
     rv = env.query("$[?search(@.a, 'a.*')]", [{ a: "the end is ac" }]);
@@ -39,13 +39,13 @@ describe("search filter", () => {
   });
   test("without caching", () => {
     const env = new JSONPathEnvironment();
-    env.filterRegister.set("search", new Search({ cacheSize: 0 }));
+    env.functionRegister.set("search", new Search({ cacheSize: 0 }));
     const rv = env.query("$[?search(@.a, 'a.*')]", [{ a: "the end is ab" }]);
     expect(rv.values()).toStrictEqual([{ a: "the end is ab" }]);
   });
   test("throw error without caching", () => {
     const env = new JSONPathEnvironment();
-    env.filterRegister.set(
+    env.functionRegister.set(
       "search",
       new Search({ cacheSize: 0, throwErrors: true }),
     );