+++ tests/path/ietf_compare.test.ts
@@ -1,5 +1,5 @@
 /**
- * Filter expression comparison test cases examples.
+ * Filter expression comparison examples.
  *
  * The test cases defined here are taken from version 20 of the JSONPath
  * internet draft, draft-ietf-jsonpath-base-20. In accordance with